chore(Writer): mv to reusable module
